Implement CallsWrappedMethod
